+++ c/src/p0009.c
@@ -19,13 +19,13 @@ Find the product abc.
 #include <inttypes.h>
 #include <stdio.h>
 
-uint64_t p0009() {
-    uint64_t answer = 0;
+uint32_t p0009() {
+    uint32_t answer = 0;
     for (uint32_t c = 3; !answer && c < 1000; c++) {
         for (uint32_t b = 2; b < c; b++) {
             uint32_t a = 1000 - c - b;
             if (a < b && a*a + b*b == c*c) {
-                answer = (uint64_t) a * b * c;
+                answer = a * b * c;
                 break;
             }
         }
@@ -35,7 +35,7 @@ uint64_t p0009() {
 
 #ifndef UNITY_END
 int main(int argc, char const *argv[]) {
-    printf("%" PRIu64 "\n", p0009());
+    printf("%" PRIu32 "\n", p0009());
     return 0;
 }
 #endif
